+++ src/CodingContract/contracts/encoding.ts
@@ -1,0 +1,219 @@
+import { getRandomIntInclusive } from "../../utils/helpers/getRandomIntInclusive";
+import { comprGenChar, comprLZDecode, comprLZEncode, comprLZGenerate } from "../../utils/CompressionContracts";
+import { type CodingContractTypes } from "../ContractTypes";
+import { HammingDecode, HammingEncode, HammingEncodeProperly } from "../../utils/HammingCodeTools";
+import { CodingContractName } from "@enums";
+
+export const encodingDefinitions = {
+  [CodingContractName.HammingCodesIntegerToEncodedBinary]: {
+    difficulty: 5,
+    desc: (n: number): string => {
+      return [
+        "You are given the following decimal value: \n",
+        `${n} \n\n`,
+        "Convert it to a binary representation and encode it as an 'extended Hamming code'.\n ",
+        "The number should be converted to a string of '0' and '1' with no leading zeroes.\n",
+        "A parity bit is inserted at position 0 and at every position N where N is a power of 2.\n",
+        "Parity bits are used to make the total number of '1' bits in a given set of data even.\n",
+        "The parity bit at position 0 considers all bits including parity bits.\n",
+        "Each parity bit at position 2^N alternately considers 2^N bits then ignores 2^N bits, starting at position 2^N.\n",
+        "The endianness of the parity bits is reversed compared to the endianness of the data bits:\n",
+        "Data bits are encoded most significant bit first and the parity bits encoded least significant bit first.\n",
+        "The parity bit at position 0 is set last.\n\n",
+        "Examples:\n\n",
+        "8 in binary is 1000, and encodes to 11110000 (pppdpddd - where p is a parity bit and d is a data bit)\n",
+        "21 in binary is 10101, and encodes to 1001101011 (pppdpdddpd)\n\n",
+        "For more information on the 'rule' of encoding, refer to Wikipedia (https://wikipedia.org/wiki/Hamming_code)",
+        "or the 3Blue1Brown videos on Hamming Codes. (https://youtube.com/watch?v=X8jsijhllIA)",
+      ].join(" ");
+    },
+    generate: (): number => {
+      const x = Math.pow(2, 4);
+      const y = Math.pow(2, getRandomIntInclusive(1, 57));
+      return getRandomIntInclusive(Math.min(x, y), Math.max(x, y));
+    },
+    solver: (data, answer) => {
+      return HammingEncode(data) === answer;
+    },
+    convertAnswer: (ans) => ans,
+    validateAnswer: (ans): ans is string => typeof ans === "string",
+  },
+  [CodingContractName.HammingCodesEncodedBinaryToInteger]: {
+    difficulty: 8,
+    desc: (n: string): string => {
+      return [
+        "You are given the following encoded binary string: \n",
+        `'${n}' \n\n`,
+        "Decode it as an 'extended Hamming code' and convert it to a decimal value.\n",
+        "The binary string may include leading zeroes.\n",
+        "A parity bit is inserted at position 0 and at every position N where N is a power of 2.\n",
+        "Parity bits are used to make the total number of '1' bits in a given set of data even.\n",
+        "The parity bit at position 0 considers all bits including parity bits.\n",
+        "Each parity bit at position 2^N alternately considers 2^N bits then ignores 2^N bits, starting at position 2^N.\n",
+        "The endianness of the parity bits is reversed compared to the endianness of the data bits:\n",
+        "Data bits are encoded most significant bit first and the parity bits encoded least significant bit first.\n",
+        "The parity bit at position 0 is set last.\n",
+        "There is a ~55% chance for an altered bit at a random index.\n",
+        "Find the possible altered bit, fix it and extract the decimal value.\n\n",
+        "Examples:\n\n",
+        "'11110000' passes the parity checks and has data bits of 1000, which is 8 in binary.\n",
+        "'1001101010' fails the parity checks and needs the last bit to be corrected to get '1001101011',",
+        "after which the data bits are found to be 10101, which is 21 in binary.\n\n",
+        "For more information on the 'rule' of encoding, refer to Wikipedia (https://wikipedia.org/wiki/Hamming_code)",
+        "or the 3Blue1Brown videos on Hamming Codes. (https://youtube.com/watch?v=X8jsijhllIA)",
+      ].join(" ");
+    },
+    generate: (): string => {
+      const _alteredBit = Math.round(Math.random());
+      const x = Math.pow(2, 4);
+      const y = Math.pow(2, getRandomIntInclusive(1, 57));
+      const _buildArray: string[] = HammingEncodeProperly(getRandomIntInclusive(Math.min(x, y), Math.max(x, y))).split(
+        "",
+      );
+      if (_alteredBit) {
+        const _randomIndex: number = getRandomIntInclusive(0, _buildArray.length - 1);
+        _buildArray[_randomIndex] = _buildArray[_randomIndex] == "0" ? "1" : "0";
+      }
+      return _buildArray.join("");
+    },
+    solver: (data, answer) => {
+      return HammingDecode(data) === answer;
+    },
+    convertAnswer: (ans) => parseInt(ans, 10),
+    validateAnswer: (ans): ans is number => typeof ans === "number",
+  },
+  [CodingContractName.CompressionIRLECompression]: {
+    difficulty: 2,
+    desc: (plaintext: string): string => {
+      return [
+        "Run-length encoding (RLE) is a data compression technique which encodes data as a series of runs of",
+        "a repeated single character. Runs are encoded as a length, followed by the character itself. Lengths",
+        "are encoded as a single ASCII digit; runs of 10 characters or more are encoded by splitting them",
+        "into multiple runs.\n\n",
+        "You are given the following input string:\n",
+        `&nbsp; &nbsp; ${plaintext}\n`,
+        "Encode it using run-length encoding with the minimum possible output length.\n\n",
+        "Examples:\n\n",
+        "&nbsp; &nbsp; aaaaabccc &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;-> &nbsp;5a1b3c\n",
+        "&nbsp; &nbsp; aAaAaA &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; -> &nbsp;1a1A1a1A1a1A\n",
+        "&nbsp; &nbsp; 111112333 &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;-> &nbsp;511233\n",
+        "&nbsp; &nbsp; zzzzzzzzzzzzzzzzzzz &nbsp;-> &nbsp;9z9z1z &nbsp;(or 9z8z2z, etc.)",
+      ].join(" ");
+    },
+    generate: (): string => {
+      const length = 50 + Math.floor(25 * (Math.random() + Math.random()));
+      let plain = "";
+
+      while (plain.length < length) {
+        const r = Math.random();
+
+        let n = 1;
+        if (r < 0.3) {
+          n = 1;
+        } else if (r < 0.6) {
+          n = 2;
+        } else if (r < 0.9) {
+          n = Math.floor(10 * Math.random());
+        } else {
+          n = 10 + Math.floor(5 * Math.random());
+        }
+
+        const c = comprGenChar();
+        plain += c.repeat(n);
+      }
+
+      return plain.substring(0, length);
+    },
+    solver: (plain, answer) => {
+      if (plain.length === 0) return answer === "";
+
+      let out = "";
+      let count = 1;
+      for (let i = 1; i < plain.length; i++) {
+        if (count < 9 && plain[i] === plain[i - 1]) {
+          count++;
+          continue;
+        }
+        out += count + plain[i - 1];
+        count = 1;
+      }
+      out += count + plain[plain.length - 1];
+      return out === answer;
+    },
+    convertAnswer: (ans) => ans.replace(/\s/g, ""),
+    validateAnswer: (ans): ans is string => typeof ans === "string",
+  },
+  [CodingContractName.CompressionIILZDecompression]: {
+    difficulty: 4,
+    desc: (compressed: string): string => {
+      return [
+        "Lempel-Ziv (LZ) compression is a data compression technique which encodes data using references to",
+        "earlier parts of the data. In this variant of LZ, data is encoded in two types of chunk. Each chunk",
+        "begins with a length L, encoded as a single ASCII digit from 1 to 9, followed by the chunk data,",
+        "which is either:\n\n",
+        "1. Exactly L characters, which are to be copied directly into the uncompressed data.\n",
+        "2. A reference to an earlier part of the uncompressed data. To do this, the length is followed",
+        "by a second ASCII digit X: each of the L output characters is a copy of the character X",
+        "places before it in the uncompressed data.\n\n",
+        "For both chunk types, a length of 0 instead means the chunk ends immediately, and the next character",
+        "is the start of a new chunk. The two chunk types alternate, starting with type 1, and the final",
+        "chunk may be of either type.\n\n",
+        "You are given the following LZ-encoded string:\n",
+        `&nbsp; &nbsp; ${compressed}\n`,
+        "Decode it and output the original string.\n\n",
+        "Example: decoding '5aaabb450723abb' chunk-by-chunk\n\n",
+        "&nbsp; &nbsp; 5aaabb &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; -> &nbsp;aaabb\n",
+        "&nbsp; &nbsp; 5aaabb45 &nbsp; &nbsp; &nbsp; &nbsp; -> &nbsp;aaabbaaab\n",
+        "&nbsp; &nbsp; 5aaabb450 &nbsp; &nbsp; &nbsp; &nbsp;-> &nbsp;aaabbaaab\n",
+        "&nbsp; &nbsp; 5aaabb45072 &nbsp; &nbsp; &nbsp;-> &nbsp;aaabbaaababababa\n",
+        "&nbsp; &nbsp; 5aaabb450723abb &nbsp;-> &nbsp;aaabbaaababababaabb",
+      ].join(" ");
+    },
+    generate: (): string => {
+      return comprLZEncode(comprLZGenerate());
+    },
+    solver: (compr, answer) => {
+      return (comprLZDecode(compr) ?? "") === answer;
+    },
+    convertAnswer: (ans) => ans.replace(/\s/g, ""),
+    validateAnswer: (ans): ans is string => typeof ans === "string",
+  },
+  [CodingContractName.CompressionIIILZCompression]: {
+    difficulty: 10,
+    desc: (plaintext: string): string => {
+      return [
+        "Lempel-Ziv (LZ) compression is a data compression technique which encodes data using references to",
+        "earlier parts of the data. In this variant of LZ, data is encoded in two types of chunk. Each chunk",
+        "begins with a length L, encoded as a single ASCII digit from 1 to 9, followed by the chunk data,",
+        "which is either:\n\n",
+        "1. Exactly L characters, which are to be copied directly into the uncompressed data.\n",
+        "2. A reference to an earlier part of the uncompressed data. To do this, the length is followed",
+        "by a second ASCII digit X: each of the L output characters is a copy of the character X",
+        "places before it in the uncompressed data.\n\n",
+        "For both chunk types, a length of 0 instead means the chunk ends immediately, and the next character",
+        "is the start of a new chunk. The two chunk types alternate, starting with type 1, and the final",
+        "chunk may be of either type.\n\n",
+        "You are given the following input string:\n",
+        `&nbsp; &nbsp; ${plaintext}\n`,
+        "Encode it using Lempel-Ziv encoding with the minimum possible output length.\n\n",
+        "Examples (some have other possible encodings of minimal length):\n",
+        "&nbsp; &nbsp; abracadabra &nbsp; &nbsp; -> &nbsp;7abracad47\n",
+        "&nbsp; &nbsp; mississippi &nbsp; &nbsp; -> &nbsp;4miss433ppi\n",
+        "&nbsp; &nbsp; aAAaAAaAaAA &nbsp; &nbsp; -> &nbsp;3aAA53035\n",
+        "&nbsp; &nbsp; 2718281828 &nbsp; &nbsp; &nbsp;-> &nbsp;627182844\n",
+        "&nbsp; &nbsp; abcdefghijk &nbsp; &nbsp; -> &nbsp;9abcdefghi02jk\n",
+        "&nbsp; &nbsp; aaaaaaaaaaaa &nbsp; &nbsp;-> &nbsp;3aaa91\n",
+        "&nbsp; &nbsp; aaaaaaaaaaaaa &nbsp; -> &nbsp;1a91031\n",
+        "&nbsp; &nbsp; aaaaaaaaaaaaaa &nbsp;-> &nbsp;1a91041",
+      ].join(" ");
+    },
+    generate: (): string => {
+      return comprLZGenerate();
+    },
+    solver: (plain, answer) => {
+      return answer.length <= comprLZEncode(plain).length && comprLZDecode(answer) === plain;
+    },
+    convertAnswer: (ans) => ans.replace(/\s/g, ""),
+    validateAnswer: (ans): ans is string => typeof ans === "string",
+  },
+} satisfies Partial<CodingContractTypes>;
